ai(danger): fix the weapon fire-cycle model first, then re-derive the danger unit on top of it
Review of 3a7a10a3 was correct on every count and this reorders the branch around it.
The overflow I reported as the headline bug was a SYMPTOM; the cause was one level down.

THE CADENCE MODEL — the ruling, derived from Armament rather than assumed.
CanFire refuses while `IsReloading || IsWaitingBurst` (Armament.cs:327) and both
counters are decremented independently in the same tick handler (:283-287), so a
weapon arming both waits the LONGER of the two, never the sum. UpdateBurst (:624-648)
arms the full BurstWait only when a burst is exhausted; UpdateMagazine (:608-622) arms
ReloadDelay only when the MAGAZINE empties, and it runs once per SHOT (:380) — so
Magazine counts shots and a magazine swap is amortised over Magazine shots.

That last point is where the two reviewing workers disagreed, and it decides it:
5.56mm.AR declares `Magazine: 100`. Its real cycle is ten 10-shot bursts of 17 ticks
plus one 150-tick swap = ~312 ticks per 20,000 damage = ~6,410, against the old
formula's 1,333. So the reviewer's "ReloadDelay dominates, 1,333 is roughly correct"
holds only if Magazine is ignored; with it, the doc worker's ~4.5x under-statement is
right and its ~6,060 reconciles. Both over-estimates (Abrams 130x, ATGM 200x) were
always right.

The old `damage x Burst x window / ReloadDelay` was therefore wrong in BOTH directions:
- ~90% of WW3MOD weapons never set ReloadDelay (the mod paces with BurstWait, which
  ArmamentInfo.RulesetLoaded:128-129 makes mandatory), hit the `reload = 1` fallback,
  and were OVER-stated by their whole cycle: TankRound.Abrams 2,300,000 vs 17,692.
- The ~14 that do set it were UNDER-stated ~4.8x, as above.
A uniform error cancels in the reference ratio. A 130x over against a 5x under does
not — it re-ranks the weapon classes, which is the one thing a threat field must get
right, and it is why the unit could not ship before this fix.

CONSEQUENTLY, AND IN THIS ORDER:
- WeaponThroughput now derives the real cycle; SustainedThroughput is split out and
  public so the model is NUnit-pinned from real weapon parameters (a warhead list is
  not constructible in a test, which is how the cadence went unexamined).
- The overflow is demoted to what it was. With the correct throughput the Abrams
  product is 5.2e7, ~41x below int.MaxValue. The longs are KEPT as a bound on
  data-driven inputs, and every comment that sold them as the fix now says otherwise.
- Test fixtures no longer hard-code `23000 * 100` as "what the field actually holds":
  they transcribe real weapon PARAMETERS from the mod yaml and compute throughput
  through the function under test, so a cadence regression moves the assertions
  instead of being ratified by them.
- The evac pin no longer calibrates against the play log's 66,834 median. That log
  predates this fix, so every heavy contact in it was stamping the clamped floor of 1
  and the distribution no longer exists. It now pins the STRUCTURAL property the loop
  violated: a tank on the cell evacuates, a decayed contact at max range does not.

THRESHOLD VALUES ARE PROVISIONAL and say so in ai.yaml. They are derived from kernel
geometry, not measured. `[danger] reference` now also reports the contributing-type
count and the min/max intensity spread, so one ordinary session says whether the
median is a fair reference and whether 50 sits above ambient and below a real contact.

TWO MORE OVERFLOW SITES, both found by review, both now pinned RED-on-revert:
- MissionCommitmentMath.DangerSpiked: `commitDanger * spikePct` wrapped negative, and
  a negative margin always loses to the floor — the percentage arm silently vanished
  and the trigger degenerated to floor-only.
- TransportDropSiteMath.ScoreDrop: `danger * dangerWeight` wrapped back into the
  POSITIVE range, and since every term is a penalty subtracted from zero that made it
  a BONUS — the picker preferred the most dangerous drop site. Live under
  RiskWeightedDropSite. Sign fixed; the weight IMBALANCE is left to a measured retune
  and recorded in discovered.md.

CORRECTION TO 3a7a10a3's COMMIT MESSAGE: it claimed FieldLoader hard-errors on an
unknown YAML key, making an old field name a mod load failure. FALSE. FieldLoader.Load
(:560-604) iterates the TYPE's fields and looks each up in the yaml; it never
enumerates yaml keys, so a stale key is silently ignored and the field keeps its C#
default. The real enforcement is Lint/CheckUnknownTraitFields.cs:85-86, under
`make test`. No straggler keys exist, so nothing is broken. Corrected in DISCOVERIES.

@STABLE DRIFT SURFACED: GroundDangerSafeUnits (supply + both offense twins) and
AirDangerSpikeUnits (both heli twins) were changing through the C# default with
nothing in ai.yaml to show it. Now written out explicitly in every block.

Also: DOCS/reference/influence-stack.md corrected (old field names, 40/120 values,
and a dated curation note on the new unit); GarrisonBotModule.MinBelievedDanger
documented as a presence test BY CONFIGURATION, not by construction — setting it to 50
would silently make it an unconverted level test; two overstated diagnoses toned to
what the code supports (MILD-unreachable, not "all in the top bucket"; frontline drops,
not "every drop").

Build clean, 1274/1274. Five pins verified RED by reverting their fix individually.
Benchmark baseline must still be re-taken: this changes @stable's danger field
unconditionally, and by MORE than 3a7a10a3 did.

+++ engine/OpenRA.Mods.Common/Traits/BotModules/PoiGoalGuard.cs
@@ -147,7 +147,12 @@ namespace OpenRA.Mods.Common.Traits
 			if (currentDanger <= commitDanger)
 				return false;
 
-			var margin = commitDanger * Math.Max(0, spikePct) / 100;
+			// long: commitDanger is a RAW danger-field reading, which on WW3MOD's scale reaches 10^7 in a
+			// contested sector, and multiplying that by a percentage overflows int and wraps NEGATIVE — which
+			// silently deletes the percentage arm (a negative margin always loses to the floor) and degenerates
+			// this trigger to floor-only, i.e. abandoning missions on ambient jitter over already-dangerous
+			// ground, which is precisely what the percentage exists to prevent.
+			var margin = (long)commitDanger * Math.Max(0, spikePct) / 100;
 			if (margin < spikeFloor)
 				margin = spikeFloor;
